+++ object_detection/detic/datasets_detic/imagenet21k_wordnet_ids.txt
