+++ object_detection/detic/datasets_detic/in21k_categories.py
