+++ object_detection/detic/datasets_detic/lvis_v1_categories.py
